Using SAVEPOINTS with Semaphore#attempt_creation to address the unstable transaction issue
Problem Statement:
When in a transaction and any of the SQL crosses paths with a database contraint that it
fails, the transaction immediately enters an unstable/unusable state.  See rails documentation
for ActiveRecord::Transactions::ClassMethods#Exception handling and rollback.

Semaphore#attempt_creation works on the theory that it can insert into the Semaphore
table and rescue on "key"conflict.  This does not appear to be correct, both based
on the docutmentation and the shere number of ActiveRecord::StatementInvalid's in my
logs.

This change simply wraps the "offending" create! via:
ActiveRecord::Base.transaction(request_new: true) do
  Semaphore.create!(...)
  true # it worked with any of our help
rescue Activerecord::RecordNotUnique
  solid_queue code
end

There are 2 open questions:
1) Is this the rollback to the SAVEPOINT working?
2) Is ActiveRecord::RecordNotUnqiue triggering the rollback or do we have handle that
   ourselves?